Order the controllers.
