+++ config/schema/og.schema.yml
@@ -20,6 +20,28 @@ field.field_settings.og_membership_reference:
       type: boolean
       label: 'Access Override'
 
+field.storage_settings.og_standard_reference:
+  type: mapping
+  label: 'Organic Groups reference field storage settings'
+  mapping:
+    target_type:
+      type: string
+      label: 'Type of entity to reference'
+
+field.field_settings.og_standard_reference:
+  type: mapping
+  label: 'Organic Groups reference field settings'
+  mapping:
+    handler:
+      type: string
+      label: 'Reference method'
+    handler_settings:
+      type: entity_reference_selection.[%parent.handler]
+      label: 'Organic Groups reference selection plugin settings'
+    access_override:
+      type: boolean
+      label: 'Access Override'
+
 og.settings:
   type: config_object
   label: 'Organic Groups settings'
