+++ src/Plugin/Field/FieldType/OgMembershipReferenceItem.php
@@ -7,8 +7,8 @@
 
 namespace Drupal\og\Plugin\Field\FieldType;
 
+use Drupal\Core\Field\FieldStorageDefinitionInterface;
 use Drupal\Core\Field\Plugin\Field\FieldType\EntityReferenceItem;
-use Drupal\Core\Form\FormStateInterface;
 
 /**
  * Class OgMembershipReferenceItem.
@@ -25,33 +25,16 @@ use Drupal\Core\Form\FormStateInterface;
  *   constraints = {"ValidReference" = {}, "ValidOgMembershipReference" = {}}
  * )
  */
-class OgMembershipReferenceItem extends EntityReferenceItem {
+class OgMembershipReferenceItem extends OgStandardReferenceItem {
 
   /**
    * {@inheritdoc}
    */
-  public static function defaultFieldSettings() {
-    $settings = parent::defaultFieldSettings();
-    $settings['access_override'] = FALSE;
-
-    return $settings;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function fieldSettingsForm(array $form, FormStateInterface $form_state) {
-    $form = parent::fieldSettingsForm($form, $form_state);
-
-    // Field access settings.
-    $form['access_override'] = [
-      '#title' => $this->t('Allow entity access to control field access'),
-      '#description' => $this->t('By default, the <em>administer group</em> permission is required to directly edit this field. Selecting this option will allow access to anybody with access to edit the entity.'),
-      '#type' => 'checkbox',
-      '#default_value' => $this->getSetting('access_override'),
-    ];
-
-    return $form;
+  public static function schema(FieldStorageDefinitionInterface $field_definition) {
+    // @todo When the FieldStorageConfig::hasCustomStorage method can be changed
+    // this will not be needed to prevent errors. Can just be an empty array,
+    // similar to PathItem.
+    return ['columns' => []];
   }
 
 }
